Add support for attaching patches on Bugzilla 4.2+
attachment.cgi requires a token field for Bugzilla 4.2 and later.
(See https://bugzilla.mozilla.org/show_bug.cgi?id=703983)
This token is not the same as the overall bug's token. So from
we need to load attachment.cgi as a GET before every POST, which
is a bit of a pain.

https://bugzilla.gnome.org/show_bug.cgi?id=700502