Restore code to remember redirects now we are using urllib2
At times, we can make many separate requests. If we start off with
a URL that bugzilla always redirects (say a HTTP URL), it's better to
"get the message" and start using the target of the redirects.

The code that used to do this was removed with the switch to urllib2.
Add back equivalent code - it can be simplified at this point because
bugzilla was changed to use 301 permanent redirects for HTTP => HTTPS
redirects, and left using 302 requests for attachment
redirects.

https://bugzilla.gnome.org/show_bug.cgi?id=725730